Fix: assert when removing element from empty array

+++ source/blender/nodes/NOD_socket_items_ops.hh
@@ -84,10 +84,11 @@ inline void remove_item(wmOperatorType *ot,
     PointerRNA node_ptr = get_active_node_to_operate_on(C, Accessor::node_type);
     bNode &node = *static_cast<bNode *>(node_ptr.data);
     SocketItemsRef ref = Accessor::get_items_from_node(node);
+    if (*ref.items_num > 0) {
     dna::array::remove_index(
         ref.items, ref.items_num, ref.active_index, *ref.active_index, Accessor::destruct_item);
-
     update_after_node_change(C, node_ptr);
+    }
     return OPERATOR_FINISHED;
   };
 }
